+++ kvmagent/kvmagent/plugins/vm_plugin.py
@@ -523,26 +523,6 @@ class VirtioCeph(object):
         e(disk, 'target', None, {'dev':'vd%s' % self.dev_letter, 'bus':'virtio'})
         return disk
 
-def makesure_qemu_with_lichbd():
-    _lichbd = lichbd.lichbd_get_qemu_path()
-    _system = kvmagent.get_qemu_path()
-    need_link = True
-
-    if os.path.islink(_system):
-        link = shell.call("set -o pipefail; ls -l %s|cut -d '>' -f 2" % (_system)).strip()
-        if link == _lichbd:
-            need_link = False
-
-    if need_link:
-        rm_cmd = "rm -f %s.tmp" % (_system)
-        shell.call(rm_cmd)
-
-        ln_cmd = "ln -s %s %s.tmp" % (_lichbd, _system)
-        shell.call(ln_cmd)
-
-        mv_cmd = "mv %s.tmp %s -f" % (_system, _system)
-        shell.call(mv_cmd)
-
 class IsoFusionstor(object):
     def __init__(self):
         self.iso = None
@@ -550,7 +530,7 @@ class IsoFusionstor(object):
     def to_xmlobject(self):
         iqn = lichbd.lichbd_get_iqn() 
         port = lichbd.lichbd_get_iscsiport()
-        makesure_qemu_with_lichbd()
+        lichbd.makesure_qemu_with_lichbd()
         snap = self.iso.path.lstrip('fusionstor:').lstrip('//')
         path = self.iso.path.lstrip('fusionstor:').lstrip('//').split('@')[0]
 
@@ -581,7 +561,7 @@ class IdeFusionstor(object):
         self.dev_letter = None
 
     def to_xmlobject(self):
-        makesure_qemu_with_lichbd()
+        lichbd.makesure_qemu_with_lichbd()
         path = self.volume.installPath.lstrip('fusionstor:').lstrip('//')
         file_format = lichbd.lichbd_get_format(path)
 
@@ -597,7 +577,7 @@ class VirtioFusionstor(object):
         self.dev_letter = None
 
     def to_xmlobject(self):
-        makesure_qemu_with_lichbd()
+        lichbd.makesure_qemu_with_lichbd()
         path = self.volume.installPath.lstrip('fusionstor:').lstrip('//')
         file_format = lichbd.lichbd_get_format(path)
